+++ pro/includes/class-foogallery-pro-datasource-taxonomy-base.php
@@ -63,7 +63,7 @@ if ( ! class_exists( 'FooGallery_Pro_Datasource_Taxonomy_Base' ) ) {
             //clear any previously cached post meta for the gallery
             $previous_datasource_value = get_post_meta( $foogallery_id, FOOGALLERY_META_DATASOURCE_VALUE, true );
 
-            if ( is_array( $previous_datasource_value ) ) {
+            if ( is_array( $previous_datasource_value ) && array_key_exists( 'taxonomy', $previous_datasource_value ) ) {
                 $taxonomy = $previous_datasource_value['taxonomy'];
                 $cache_post_meta_key = FOOGALLERY_META_DATASOURCE_CACHED_ATTACHMENTS . '_' . $taxonomy;
                 delete_post_meta($foogallery_id, $cache_post_meta_key);
